IdentityProvider/requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
deta==1.1.0
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
pydantic==1.9.1
PyJWT==2.4.0
python-decouple==3.6
requests==2.28.1
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.12
uvicorn==0.17.6